feat: relaciona user com task

+++ src/task/task.service.ts
@@ -7,17 +7,23 @@ import { v4 as uuid } from 'uuid';
 export class TaskService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(task: TaskDTO) {
+  async create(task: TaskDTO, userId: string) {
     task.id = uuid();
     task.status = TaskStatusEnum.PENDING;
     const response = await this.prisma.task.create({
-      data: { ...task, due_date: new Date(task.due_date).toISOString() },
+      data: {
+        ...task,
+        due_date: new Date(task.due_date).toISOString(),
+        user_id: userId,
+      },
     });
     return response;
   }
 
-  async findById(id: string) {
-    const foundTask = await this.prisma.task.findMany({ where: { id } });
+  async findById(id: string, user_id: string) {
+    const foundTask = await this.prisma.task.findMany({
+      where: { id, user_id },
+    });
 
     if (foundTask.length >= 0) {
       return foundTask;
@@ -25,19 +31,19 @@ export class TaskService {
     throw new NotFoundException('Task not found');
   }
 
-  async findAll() {
+  async findAll(user_id: string) {
     try {
-      const foundTask = await this.prisma.task.findMany();
+      const foundTask = await this.prisma.task.findMany({ where: { user_id } });
       return foundTask;
     } catch {
       throw new NotFoundException('Tasks not found');
     }
   }
 
-  async update(id: string, task: TaskDTO) {
+  async update(id: string, task: TaskDTO, user_id: string) {
     try {
       await this.prisma.task.update({
-        where: { id },
+        where: { id, user_id },
         data: {
           ...task,
           due_date: new Date(task.due_date).toISOString(),
@@ -48,9 +54,9 @@ export class TaskService {
     }
   }
 
-  async delete(id: string) {
+  async delete(id: string, user_id: string) {
     try {
-      await this.prisma.task.delete({ where: { id } });
+      await this.prisma.task.delete({ where: { id, user_id } });
     } catch {
       throw new NotFoundException('Task not found');
     }
